refactor: Move hairstyleData.ts from assets to constants folder
- Move hairstyleData.ts from assets/images to constants (where it belongs)
- Update import path in home screen component
- Fix avatar image path to use correct relative import
- TypeScript files should never be in assets/images directory
